broken to components and hook Table.tsx
